# Base ScanFilterAndProjectOperator on WorkProcessors