--- a/presto-main/src/main/java/io/prestosql/operator/WorkProcessor.java
+++ b/presto-main/src/main/java/io/prestosql/operator/WorkProcessor.java
@@ -16,6 +16,7 @@ package io.prestosql.operator;
 import com.google.common.collect.Iterators;
 import com.google.common.util.concurrent.ListenableFuture;
 
+import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 
 import java.util.Comparator;
@@ -191,9 +192,9 @@ public interface WorkProcessor<T>
     @Immutable
     final class TransformationState<T>
     {
-        private static final TransformationState<?> NEEDS_MORE_DATE_STATE = new TransformationState<>(Type.NEEDS_MORE_DATA, true, Optional.empty(), Optional.empty());
-        private static final TransformationState<?> YIELD_STATE = new TransformationState<>(Type.YIELD, false, Optional.empty(), Optional.empty());
-        private static final TransformationState<?> FINISHED_STATE = new TransformationState<>(Type.FINISHED, false, Optional.empty(), Optional.empty());
+        private static final TransformationState<?> NEEDS_MORE_DATE_STATE = new TransformationState<>(Type.NEEDS_MORE_DATA, true, null, null);
+        private static final TransformationState<?> YIELD_STATE = new TransformationState<>(Type.YIELD, false, null, null);
+        private static final TransformationState<?> FINISHED_STATE = new TransformationState<>(Type.FINISHED, false, null, null);
 
         enum Type
         {
@@ -206,15 +207,17 @@ public interface WorkProcessor<T>
 
         private final Type type;
         private final boolean needsMoreData;
-        private final Optional<T> result;
-        private final Optional<ListenableFuture<?>> blocked;
+        @Nullable
+        private final T result;
+        @Nullable
+        private final ListenableFuture<?> blocked;
 
-        private TransformationState(Type type, boolean needsMoreData, Optional<T> result, Optional<ListenableFuture<?>> blocked)
+        private TransformationState(Type type, boolean needsMoreData, @Nullable T result, @Nullable ListenableFuture<?> blocked)
         {
             this.type = requireNonNull(type, "type is null");
             this.needsMoreData = needsMoreData;
-            this.result = requireNonNull(result, "result is null");
-            this.blocked = requireNonNull(blocked, "blocked is null");
+            this.result = result;
+            this.blocked = blocked;
         }
 
         /**
@@ -234,7 +237,7 @@ public interface WorkProcessor<T>
          */
         public static <T> TransformationState<T> blocked(ListenableFuture<?> blocked)
         {
-            return new TransformationState<>(Type.BLOCKED, false, Optional.empty(), Optional.of(blocked));
+            return new TransformationState<>(Type.BLOCKED, false, null, requireNonNull(blocked, "blocked is null"));
         }
 
         /**
@@ -262,7 +265,7 @@ public interface WorkProcessor<T>
          */
         public static <T> TransformationState<T> ofResult(T result, boolean needsMoreData)
         {
-            return new TransformationState<>(Type.RESULT, needsMoreData, Optional.of(result), Optional.empty());
+            return new TransformationState<>(Type.RESULT, needsMoreData, requireNonNull(result, "result is null"), null);
         }
 
         /**
@@ -284,12 +287,14 @@ public interface WorkProcessor<T>
             return needsMoreData;
         }
 
-        Optional<T> getResult()
+        @Nullable
+        T getResult()
         {
             return result;
         }
 
-        Optional<ListenableFuture<?>> getBlocked()
+        @Nullable
+        ListenableFuture<?> getBlocked()
         {
             return blocked;
         }
@@ -298,8 +303,8 @@ public interface WorkProcessor<T>
     @Immutable
     final class ProcessState<T>
     {
-        private static final ProcessState<?> YIELD_STATE = new ProcessState<>(Type.YIELD, Optional.empty(), Optional.empty());
-        private static final ProcessState<?> FINISHED_STATE = new ProcessState<>(Type.FINISHED, Optional.empty(), Optional.empty());
+        private static final ProcessState<?> YIELD_STATE = new ProcessState<>(Type.YIELD, null, null);
+        private static final ProcessState<?> FINISHED_STATE = new ProcessState<>(Type.FINISHED, null, null);
 
         public enum Type
         {
@@ -310,14 +315,16 @@ public interface WorkProcessor<T>
         }
 
         private final Type type;
-        private final Optional<T> result;
-        private final Optional<ListenableFuture<?>> blocked;
+        @Nullable
+        private final T result;
+        @Nullable
+        private final ListenableFuture<?> blocked;
 
-        private ProcessState(Type type, Optional<T> result, Optional<ListenableFuture<?>> blocked)
+        private ProcessState(Type type, @Nullable T result, @Nullable ListenableFuture<?> blocked)
         {
             this.type = requireNonNull(type, "type is null");
-            this.result = requireNonNull(result, "result is null");
-            this.blocked = requireNonNull(blocked, "blocked is null");
+            this.result = result;
+            this.blocked = blocked;
         }
 
         /**
@@ -325,7 +332,7 @@ public interface WorkProcessor<T>
          */
         public static <T> ProcessState<T> blocked(ListenableFuture<?> blocked)
         {
-            return new ProcessState<>(Type.BLOCKED, Optional.empty(), Optional.of(blocked));
+            return new ProcessState<>(Type.BLOCKED, null, requireNonNull(blocked, "blocked is null"));
         }
 
         /**
@@ -342,7 +349,7 @@ public interface WorkProcessor<T>
          */
         public static <T> ProcessState<T> ofResult(T result)
         {
-            return new ProcessState<>(Type.RESULT, Optional.of(result), Optional.empty());
+            return new ProcessState<>(Type.RESULT, requireNonNull(result, "result is null"), null);
         }
 
         /**
@@ -359,12 +366,14 @@ public interface WorkProcessor<T>
             return type;
         }
 
-        public Optional<T> getResult()
+        @Nullable
+        public T getResult()
         {
             return result;
         }
 
-        public Optional<ListenableFuture<?>> getBlocked()
+        @Nullable
+        public ListenableFuture<?> getBlocked()
         {
             return blocked;
         }