--- a/presto-main/src/main/java/io/prestosql/operator/WorkProcessorUtils.java
+++ b/presto-main/src/main/java/io/prestosql/operator/WorkProcessorUtils.java
@@ -301,11 +301,11 @@ public final class WorkProcessorUtils
                         case NEEDS_MORE_DATA:
                             break;
                         case BLOCKED:
-                            return ProcessState.blocked(state.getBlocked().get());
+                            return ProcessState.blocked(state.getBlocked());
                         case YIELD:
                             return ProcessState.yield();
                         case RESULT:
-                            return ProcessState.ofResult(state.getResult().get());
+                            return ProcessState.ofResult(state.getResult());
                         case FINISHED:
                             return ProcessState.finished();
                     }
@@ -353,14 +353,14 @@ public final class WorkProcessorUtils
         @Override
         public boolean isBlocked()
         {
-            return state != null && state.getType() == ProcessState.Type.BLOCKED && !state.getBlocked().get().isDone();
+            return state != null && state.getType() == ProcessState.Type.BLOCKED && !state.getBlocked().isDone();
         }
 
         @Override
         public ListenableFuture<?> getBlockedFuture()
         {
             checkState(state != null && state.getType() == ProcessState.Type.BLOCKED, "Must be blocked to get blocked future");
-            return state.getBlocked().get();
+            return state.getBlocked();
         }
 
         @Override
@@ -373,7 +373,7 @@ public final class WorkProcessorUtils
         public T getResult()
         {
             checkState(state != null && state.getType() == ProcessState.Type.RESULT, "process() must return true and must not be finished");
-            return state.getResult().get();
+            return state.getResult();
         }
     }