Part of: https://github.com/prestosql/presto/issues/49

This paves way for:
1. Late source pages materialization.
2. Reusing pages source (e.g: ORC) resources between pages.